Add visualizations to status page
This adds both a view of current GitHub actions job status (# of queued,
pending, and running jobs) and a heatmap that shows how job failures (based on the jobs from
pytorch-master) correlate with each other. The first is useful to see when GitHub is
backed up on queueing, and the second for identifying jobs that commonly
fail together (meaning we can probably get rid of one).

[ghstack-poisoned]