Add logs + statuses to PR page
This adds a log viewer for GHA job logs, which are downloaded on-demand per job when opened with the chevron in the attached image. I haven't used it a ton but it could be a good alternative in cases where the GitHub log viewer is too slow. This also adds checks/Xs for job statuses.