[pr page] Add test report renderer
This adds a button to download and render failures from a test report zip. It only shows errors/failures (we could add some more info in a follow up, but this is probably what is most pressing). This brings in non-react depdenencies of fflate (which I found to be much faster than JSZip) and highlight.js (which right now does its best on the traceback, we could add some better parsing on this in the future).