Group related jobs into a single column
This adds a bunch of groups of jobs (based on a regex that is matched against the name) that makes it so the HUD is compressed enough to show up without any horizontal scrolling on a reasonable screen. When you expand a job the jobs are out of sorted order since they are inserted right after the group's name. There is also another option to disable the grouping altogether. See the deploy preview for usage.

The actual groups themselves are also totally open for debate / bikeshedding. A follow up diff will merge in the binary builds and fix #68.

This also updates `rc-tooltip` since we were already using an API that wasn't present in the old version

[ghstack-poisoned]