Move updating code to GHA
This moves the AWS Lambda from test-infra to GitHub Actions. This has a few advantages, mainly that everything to run the HUD is now co-located in one repo under a familiar interface that doesn't require digging through the AWS console. The HUD could also track its own status. GitHub API requests appear to be way faster on the GitHub runners as well which is a nice bonus.

This plan is to land this, make sure it works for a few days, then change the prefix from `v5` to `v6` in the viewing code and make sure that works for a few days, then we can disable and delete the AWS lambda job.